# contracts/agent-runtime/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Runtime HTTP API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /agent-manifests:
    get:
      operationId: listAgentManifests
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Active Agent Manifest catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifestPage'
        default:
          $ref: '#/components/responses/Error'
  /agent-manifests/{manifestID}:
    get:
      operationId: getAgentManifest
      parameters:
        - $ref: '#/components/parameters/AgentManifestID'
        - $ref: '#/components/parameters/OptionalRevision'
      responses:
        '200':
          description: One immutable Agent Manifest revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifest'
        default:
          $ref: '#/components/responses/Error'
...
  /runs/{runID}/handoff-joins:
    post:
      operationId: createRunHandoffJoin
      parameters:
        - $ref: '#/components/parameters/RunID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunHandoffJoinRequest'
      responses:
        '201':
          description: Created durable handoff fan-in contract and paused the parent run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunHandoffJoin'
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: listRunHandoffJoins
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/RunHandoffJoinStatus'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Durable handoff fan-in contracts for the parent run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunHandoffJoinPage'
        default:
          $ref: '#/components/responses/Error'
  /runs/{runID}/handoff-joins/{joinID}:
    get:
      operationId: getRunHandoffJoin
      parameters:
        - $ref: '#/components/parameters/RunID'
        - $ref: '#/components/parameters/RunHandoffJoinID'
      responses:
        '200':
          description: One durable handoff fan-in contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunHandoffJoin'
        default:
          $ref: '#/components/responses/Error'
  /runs/{runID}/task-tree:
    get:
      operationId: getRunTaskTree
      parameters:
        - $ref: '#/components/parameters/RunID'
      responses:
        '200':
          description: Durable parent and child run task tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTaskTree'
        default:
          $ref: '#/components/responses/Error'
...
  /admin/agentruntime/agent-manifests:
    get:
      operationId: listAdminAgentManifests
      parameters:
        - $ref: '#/components/parameters/AgentManifestStatus'
        - $ref: '#/components/parameters/AgentManifestScope'
        - $ref: '#/components/parameters/AgentManifestTenantID'
        - $ref: '#/components/parameters/AgentManifestOwnerActorID'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Latest Agent Manifest revisions for administrators
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifestPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createAgentManifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentManifestRevisionRequest'
      responses:
        '201':
          description: Created Agent Manifest revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifest'
        default:
          $ref: '#/components/responses/Error'
  /admin/agentruntime/agent-manifests/{manifestID}/revisions:
    post:
      operationId: reviseAgentManifest
      parameters:
        - $ref: '#/components/parameters/AgentManifestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentManifestRevisionRequest'
      responses:
        '201':
          description: Appended Agent Manifest revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifest'
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    RunID:
      name: runID
      in: path
      required: true
      schema: {type: string}
    RunHandoffJoinID:
      name: joinID
      in: path
      required: true
      schema: {type: string}
    AgentManifestID:
      name: manifestID
      in: path
      required: true
      schema: {type: string}
...
    Offset:
      name: offset
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
    AgentManifestStatus:
      name: status
      in: query
      required: false
      schema: {type: string, enum: [active, disabled]}
    AgentManifestScope:
      name: scope
      in: query
      required: false
      schema: {type: string, enum: [actor, tenant, system]}
    AgentManifestTenantID:
      name: tenantID
      in: query
      required: false
      schema: {type: string, maxLength: 64}
    AgentManifestOwnerActorID:
      name: ownerActorID
      in: query
      required: false
      schema: {type: string, maxLength: 64}
...
    RunHandoffJoinStatus:
      name: status
      in: query
      required: false
      schema: {type: string, enum: [pending, ready, failed, cancelled]}
    OptionalRevision:
      name: revision
      in: query
      required: false
      schema: {type: string}
...
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1}
...
  responses:
    Error:
      description: Stable Runtime error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ActorRef:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        tenantID: {type: string}
        id: {type: string}
    ThreadRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
    ProjectionRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
    ResourceRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
        revision: {type: string}
    AgentManifestRevisionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        manifestID: {type: string, maxLength: 64}
        expectedRevision: {type: integer, minimum: 0}
        scope: {type: string, enum: [actor, tenant, system]}
        tenantID: {type: string, maxLength: 64}
        ownerActorID: {type: string, maxLength: 64}
        name: {type: string, minLength: 1, maxLength: 128}
        description: {type: string, maxLength: 4000}
        instructions: {type: string, maxLength: 20000}
        status: {type: string, enum: [active, disabled]}
        modelName: {type: string, maxLength: 128}
        toolKeys: {type: array, maxItems: 128, items: {type: string}}
        skillKeys: {type: array, maxItems: 128, items: {type: string}}
        maxChildRuns: {type: integer, minimum: 1, maximum: 16}
        maxDepth: {type: integer, minimum: 1, maximum: 6}
        maxLLMCalls: {type: integer, minimum: 2, maximum: 32, description: Zero or omitted inherits the environment limit.}
        maxToolCalls: {type: integer, minimum: 1, maximum: 64, description: Zero or omitted inherits the environment limit.}
        revisionNote: {type: string, maxLength: 255}
    AgentManifest:
      type: object
      additionalProperties: false
      required: [manifestID, revision, ref, scope, tenantID, ownerActorID, name, status, toolKeys, skillKeys, maxChildRuns, maxDepth, maxLLMCalls, maxToolCalls, createdAt, updatedAt]
      properties:
        manifestID: {type: string}
        revision: {type: integer, minimum: 1}
        ref:
          $ref: '#/components/schemas/ResourceRef'
        scope: {type: string, enum: [actor, tenant, system]}
        tenantID: {type: string}
        ownerActorID: {type: string}
        name: {type: string}
        description: {type: string}
        instructions: {type: string}
        status: {type: string, enum: [active, disabled]}
        modelName: {type: string}
        toolKeys: {type: array, items: {type: string}}
        skillKeys: {type: array, items: {type: string}}
        maxChildRuns: {type: integer, minimum: 1}
        maxDepth: {type: integer, minimum: 1}
        maxLLMCalls: {type: integer, minimum: 0, maximum: 32}
        maxToolCalls: {type: integer, minimum: 0, maximum: 64}
        revisionNote: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    AgentManifestPage:
      type: object
      additionalProperties: false
      required: [total, results]
      properties:
        total: {type: integer, format: int64, minimum: 0}
        results:
          type: array
          items:
            $ref: '#/components/schemas/AgentManifest'
...
    CreateRunHandoffJoinRequest:
      type: object
      additionalProperties: false
      required: [clientJoinID, handoffIDs]
      properties:
        clientJoinID: {type: string, minLength: 1, maxLength: 64}
        handoffIDs:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 64}
        mode: {type: string, enum: [all, any, quorum], default: all}
        quorum: {type: integer, minimum: 1, maximum: 16}
        failurePolicy: {type: string, enum: [collect, fail_fast], default: collect}
        timeoutSeconds: {type: integer, minimum: 60, maximum: 604800, default: 3600}
        timeoutPolicy: {type: string, enum: [cancel_pending, leave_running], default: cancel_pending}
...
    Run:
      type: object
      additionalProperties: true
      required: [schemaVersion, runtimeKind, actor, thread, runID, requestID, status]
      properties:
        schemaVersion: {type: integer, enum: [1]}
        runtimeKind: {type: string, enum: [text, plan_execute, workflow, team]}
        actor:
          $ref: '#/components/schemas/ActorRef'
        thread:
          $ref: '#/components/schemas/ThreadRef'
        runID: {type: string}
        requestID: {type: string}
        status: {type: string}
        agentManifestRef:
          $ref: '#/components/schemas/ResourceRef'
        workflowDefinitionRef:
          $ref: '#/components/schemas/ResourceRef'
        agentName: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        handoffID: {type: string}
        depth: {type: integer, minimum: 1}
...
    RunHandoff:
      type: object
      additionalProperties: false
      required: [handoffID, clientHandoffID, rootRunID, parentRunID, childRunID, agentManifest, agentName, goal, status, depth, inputProjectionRef, resultOutputIDs, createdAt, updatedAt]
      properties:
        handoffID: {type: string}
        clientHandoffID: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        childRunID: {type: string}
        agentManifest:
          $ref: '#/components/schemas/ResourceRef'
        agentName: {type: string}
        goal: {type: string}
        status: {type: string, enum: [queued, completed, failed, cancelled]}
        depth: {type: integer, minimum: 1}
        inputProjectionRef:
          $ref: '#/components/schemas/ProjectionRef'
        resultSummary: {type: string}
        resultOutputIDs: {type: array, items: {type: string}}
        errorCode: {type: string}
        errorMessage: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time, nullable: true}
    RunHandoffJoin:
      type: object
      additionalProperties: false
      required: [joinID, clientJoinID, rootRunID, parentRunID, handoffIDs, mode, quorum, failurePolicy, timeoutSeconds, timeoutPolicy, deadlineAt, status, completedCount, failedCount, cancelledCount, pendingCount, resultHandoffIDs, createdAt, updatedAt]
      properties:
        joinID: {type: string}
        clientJoinID: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        handoffIDs:
          type: array
          items: {type: string}
        mode: {type: string, enum: [all, any, quorum]}
        quorum: {type: integer, minimum: 1}
        failurePolicy: {type: string, enum: [collect, fail_fast]}
        timeoutSeconds: {type: integer, minimum: 60, maximum: 604800}
        timeoutPolicy: {type: string, enum: [cancel_pending, leave_running]}
        deadlineAt: {type: string, format: date-time}
        status: {type: string, enum: [pending, ready, failed, cancelled]}
        completedCount: {type: integer, minimum: 0}
        failedCount: {type: integer, minimum: 0}
        cancelledCount: {type: integer, minimum: 0}
        pendingCount: {type: integer, minimum: 0}
        resultHandoffIDs:
          type: array
          items: {type: string}
        errorCode: {type: string}
        errorMessage: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resolvedAt: {type: string, format: date-time, nullable: true}
    RunHandoffJoinPage:
      type: object
      additionalProperties: false
      required: [total, results]
      properties:
        total: {type: integer, format: int64, minimum: 0}
        results:
          type: array
          items:
            $ref: '#/components/schemas/RunHandoffJoin'
...
    RunTask:
      type: object
      additionalProperties: false
      required: [handoff, run]
      properties:
        handoff:
          $ref: '#/components/schemas/RunHandoff'
        run:
          $ref: '#/components/schemas/Run'
    RunTaskTree:
      type: object
      additionalProperties: false
      required: [rootRunID, currentRunID, rootRun, tasks, joins]
      properties:
        rootRunID: {type: string}
        currentRunID: {type: string}
        rootRun:
          $ref: '#/components/schemas/Run'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/RunTask'
        joins:
          type: array
          items:
            $ref: '#/components/schemas/RunHandoffJoin'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestID]
          properties:
            code: {type: string}
            message: {type: string}
            requestID: {type: string}